feat(aws_cli): add optional argument to install

+++ deployments/aws_s3.sh
@@ -6,15 +6,19 @@
 # * AWS_SECRET_ACCESS_KEY
 # * AWS_DEFAULT_REGION
 # * AWS_S3_BUCKET
+# * AWS_CLI_RUN_INSTALL - Set a value if you wanna install the AWS-CLI also.
 #
 # Include in your builds via
 # \curl -sSL https://raw.githubusercontent.com/codeship/scripts/master/deployments/aws_s3.sh | bash -s
 
 # Install AWS-CLI Latest without sudo. (bundled installer - http://docs.aws.amazon.com/cli/latest/userguide/installing.html)
-curl "https://s3.amazonaws.com/aws-cli/awscli-bundle.zip" -o "awscli-bundle.zip"
-unzip awscli-bundle.zip
-./awscli-bundle/install -b ~/bin/aws
-rm -rf awscli-bundle
+if [ -n "${AWS_CLI_RUN_INSTALL}" ]; then # Checks if not empty
+  echo "Installing AWS-CLI using bundled installer"
+  curl "https://s3.amazonaws.com/aws-cli/awscli-bundle.zip" -o "awscli-bundle.zip"
+	unzip awscli-bundle.zip
+	./awscli-bundle/install -b ~/bin/aws
+	rm -rf awscli-bundle
+fi
 
 AWS_ACCESS_KEY_ID=${AWS_ACCESS_KEY_ID:?'You need to configure the AWS_ACCESS_KEY_ID environment variable!'}
 AWS_SECRET_ACCESS_KEY=${AWS_SECRET_ACCESS_KEY:?'You need to configure the AWS_SECRET_ACCESS_KEY environment variable!'}
